chore: clean up a few things
