+++ integration-tests/src/sbom/test/mod.rs
@@ -1,23 +1,28 @@
 #![cfg(test)]
 
+mod basic;
 mod perf;
 
 use lzma::LzmaReader;
+use spdx_rs::models::SPDX;
+use std::collections::HashSet;
 use std::fs::File;
+use std::future::Future;
 use std::io::{BufReader, Read};
 use std::path::PathBuf;
 use std::str::FromStr;
 use std::time::Instant;
-use test_context::test_context;
-use test_log::test;
-use tracing::{info_span, instrument};
+use tracing::{info_span, instrument, Instrument};
 use trustify_common::db::test::TrustifyContext;
-use trustify_common::db::Transactional;
-use trustify_entity::relationship::Relationship;
-use trustify_module_fetch::model::sbom::SbomPackage;
-use trustify_module_fetch::service::sbom::Which;
+use trustify_common::db::Database;
 use trustify_module_fetch::service::FetchService;
-use trustify_module_ingestor::graph::{sbom::spdx::parse_spdx, sbom::spdx::Information, Graph};
+use trustify_module_ingestor::graph::{
+    sbom::{
+        spdx::{parse_spdx, Information},
+        SbomContext,
+    },
+    Graph,
+};
 
 #[instrument]
 pub fn open_sbom(name: &str) -> anyhow::Result<impl Read> {
@@ -33,99 +38,69 @@ pub fn open_sbom_xz(name: &str) -> anyhow::Result<impl Read> {
     Ok(LzmaReader::new_decompressor(open_sbom(name)?)?)
 }
 
-#[test_context(TrustifyContext, skip_teardown)]
-#[instrument]
-#[test(tokio::test)]
-async fn parse_spdx_quarkus(ctx: TrustifyContext) -> Result<(), anyhow::Error> {
-    let db = ctx.db;
-    let system = Graph::new(db.clone());
-    let fetch = FetchService::new(db);
+/// remove all relationships having broken references
+pub fn fix_rels(mut spdx: SPDX) -> SPDX {
+    let mut ids = spdx
+        .package_information
+        .iter()
+        .map(|p| &p.package_spdx_identifier)
+        .collect::<HashSet<_>>();
 
-    // nope, has bad license expressions
-    let sbom_data = open_sbom("quarkus-bom-2.13.8.Final-redhat-00004.json")?;
+    ids.insert(&spdx.document_creation_information.spdx_identifier);
 
-    let start = Instant::now();
-    let parse_time = start.elapsed();
+    spdx.relationships.retain(|rel| {
+        let r = ids.contains(&rel.spdx_element_id) && ids.contains(&rel.related_spdx_element);
+        if !r {
+            log::warn!(
+                "Dropping - left: {}, rel: {:?}, right: {}",
+                rel.spdx_element_id,
+                rel.relationship_type,
+                rel.related_spdx_element
+            );
+        }
+        r
+    });
 
-    let (spdx, _) = info_span!("parse json").in_scope(|| parse_spdx(sbom_data))?;
-
-    let start = Instant::now();
-    let tx = system.transaction().await?;
-
-    let sbom = system
-        .ingest_sbom(
-            "test.com/my-sbom.json",
-            "10",
-            &spdx.document_creation_information.spdx_document_namespace,
-            Information(&spdx),
-            &tx,
-        )
-        .await?;
-
-    sbom.ingest_spdx(spdx, &tx).await?;
-    let ingest_time = start.elapsed();
-    let start = Instant::now();
-
-    // commit, then test
-    tx.commit().await?;
-
-    let described = fetch
-        .describes_packages(sbom.sbom.sbom_id, Default::default(), Transactional::None)
-        .await?;
-    log::info!("{:#?}", described);
-    assert_eq!(1, described.items.len());
-    let first = &described.items[0];
-    assert_eq!(
-        &SbomPackage {
-            id: "SPDXRef-b52acd7c-3a3f-441e-aef0-bbdaa1ec8acf".into(),
-            name: "quarkus-bom".into(),
-            purl: vec![
-                "pkg://maven/com.redhat.quarkus.platform/quarkus-bom@2.13.8.Final-redhat-00004?repository_url=https://maven.repository.redhat.com/ga/&type=pom".into()
-            ],
-            cpe: vec!["cpe:/a:redhat:quarkus:2.13:*:el8:*".to_string()],
-        },
-        first
-    );
-
-    let contains = fetch
-        .related_packages(
-            sbom.sbom.sbom_id,
-            Relationship::ContainedBy,
-            first,
-            Transactional::None,
-        )
-        .await?;
-
-    log::info!("{}", contains.len());
-
-    assert!(contains.len() > 500);
-
-    let query_time = start.elapsed();
-
-    log::info!("parse {}ms", parse_time.as_millis());
-    log::info!("ingest {}ms", ingest_time.as_millis());
-    log::info!("query {}ms", query_time.as_millis());
-
-    Ok(())
+    spdx
 }
 
-#[test_context(TrustifyContext, skip_teardown)]
-#[test(tokio::test)]
-async fn test_parse_spdx(ctx: TrustifyContext) -> Result<(), anyhow::Error> {
+pub struct WithContext {
+    pub sbom: SbomContext,
+    pub db: Database,
+    pub graph: Graph,
+    pub fetch: FetchService,
+}
+
+#[instrument(skip(ctx, f))]
+async fn test_with<F, Fut>(ctx: TrustifyContext, sbom: &str, f: F) -> anyhow::Result<()>
+where
+    F: FnOnce(WithContext) -> Fut,
+    Fut: Future<Output = anyhow::Result<()>>,
+{
+    // The `ctx` must live until the end of this function. Otherwise, it will tear down the database
+    // while we're testing. So we take the `db` and offer it to the test, but we hold on the `ctx`
+    // instance until that test returns.
+
     let db = ctx.db;
-    let system = Graph::new(db.clone());
-    let fetch = FetchService::new(db);
-
-    let sbom = open_sbom("ubi9-9.2-755.1697625012.json")?;
-
-    let tx = system.transaction().await?;
+    let graph = Graph::new(db.clone());
+    let fetch = FetchService::new(db.clone());
 
     let start = Instant::now();
-    let (spdx, _) = parse_spdx(sbom)?;
+    let (spdx, _) = info_span!("parse json").in_scope(|| {
+        Ok::<_, anyhow::Error>(if sbom.ends_with(".xz") {
+            parse_spdx(open_sbom_xz(sbom)?)?
+        } else {
+            parse_spdx(open_sbom(sbom)?)?
+        })
+    })?;
+
+    let spdx = fix_rels(spdx);
     let parse_time = start.elapsed();
 
+    let tx = graph.transaction().await?;
+
     let start = Instant::now();
-    let sbom = system
+    let sbom = graph
         .ingest_sbom(
             "test.com/my-sbom.json",
             "10",
@@ -134,43 +109,38 @@ async fn test_parse_spdx(ctx: TrustifyContext) -> Result<(), anyhow::Error> {
             &tx,
         )
         .await?;
+    let ingest_time_1 = start.elapsed();
 
-    sbom.ingest_spdx(spdx, &tx).await?;
-
-    tx.commit().await?;
-
-    let ingest_time = start.elapsed();
     let start = Instant::now();
+    sbom.ingest_spdx(spdx.clone(), &tx).await?;
+    let ingest_time_2 = start.elapsed();
 
-    let described = fetch
-        .describes_packages(sbom.sbom.sbom_id, Default::default(), Transactional::None)
-        .await?;
+    // commit
 
-    assert_eq!(1, described.total);
-    let first = &described.items[0];
+    let start = Instant::now();
+    tx.commit().await?;
+    let commit_time = start.elapsed();
 
-    let contains = fetch
-        .fetch_related_packages(
-            sbom.sbom.sbom_id,
-            Default::default(),
-            Default::default(),
-            Which::Right,
-            first,
-            Some(Relationship::ContainedBy),
-            (),
-        )
-        .await?
-        .items;
+    // now test
 
-    log::info!("{}", contains.len());
+    let start = Instant::now();
+    f(WithContext {
+        sbom,
+        db,
+        graph,
+        fetch,
+    })
+    .instrument(info_span!("assert"))
+    .await?;
+    let test_time = start.elapsed();
 
-    assert!(contains.len() > 500);
+    // log durations
 
-    let query_time = start.elapsed();
-
-    log::info!("parse {}ms", parse_time.as_millis());
-    log::info!("ingest {}ms", ingest_time.as_millis());
-    log::info!("query {}ms", query_time.as_millis());
+    log::info!("parse: {}", humantime::Duration::from(parse_time));
+    log::info!("ingest 1: {}", humantime::Duration::from(ingest_time_1));
+    log::info!("ingest 2: {}", humantime::Duration::from(ingest_time_2));
+    log::info!("commit: {}", humantime::Duration::from(commit_time));
+    log::info!("test: {}", humantime::Duration::from(test_time));
 
     Ok(())
 }
